Kereső és szűrő frontend kialakítása (újra gondolva, commitot töröltem)
