--- client/src/components/SearchBar.jsx
@@ -1,9 +1,0 @@
-function SearchBar() {
-    return (
-      <div className="flex items-center mb-4 p-4 lg:px-8 lg:py-5 rounded w-full">
-        <input type="text" id="search" className="border-0 grow p-3" placeholder="Search" />
-      </div>
-    );
-  }
-  
-  export default SearchBar;
